Python/FilesUpload/requirements.txt

streamlit
openai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_openai
langchain-community
tiktoken
typing_extensions


